pyproject.toml
[tool.poetry]
name = "image-search"
version = "0.1.0"
authors = ["Kris Urbanczyk"]
readme = "README.md"
description = "image search engine by feature"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "~0.103.1"
uvicorn = "~0.23.2"
mysql-connector-python = "~8.1.0"
sqlalchemy = "~2.0.0"
alembic = "~1.11.1"
taskiq-redis = "^1.0.0"
taskiq-aio-pika = "^0.4.0"
aio-pika = "^9.4.3"
taskiq = "^0.11.6"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
asyncmy = "^0.2.9"
pytest-asyncio = "^0.23.8"
pillow = "^10.4.0"
python-multipart = "^0.0.9"
redis = "^5.0.8"
opencv-python = "^4.10.0.84"
numpy = "^2.0.1"
filelock = "^3.15.4"
cryptography = "^43.0.0"
sqlalchemy2-stubs = "^0.0.2a38"
opencv-stubs = "^0.0.10"
types-redis = "^4.6.0.20240806"


[tool.poetry.group.dev.dependencies]
pytest = "~8.3.2"
black = "~24.8.0"
isort = "~5.13.2"
mypy = "~1.11.1"
ruff = "~0.5.7"
pytest-env = "^1.1.3"
polyfactory = "^2.16.2"
opencv-stubs = "^0.0.10"
httpx = "^0.27.0"
coverage = "^7.6.1"


[tool.black]
target-version = ["py311"]
line-length = 120

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
create_redisearch_indices = "image_search.image.vector_search_methods.main:create_redisearch_indices"